Update README and fix bug in qr-by-image-size
Update README wtih descriptions and pictures of the new argument.

Update existing readme examples with new argument.

Up module size to 16 pixels and remove image size validation on the QR by image size. (We're going to resize the image in a moment anyway, so the validation doesn't make much sense.)

The reason we need a larger module size is because you can only do square modules if each module is 1px*1px. If we want the module drawers to be available, we need a larger pixel size.